- fixed location reporting - fixed some parsing issues - still has trouble when "." is the first character in a query

+++ include/aejson/loc.h
@@ -14,10 +14,10 @@
  */
 typedef struct aejson_loc
 {
-     size_t first_line;
-     size_t first_column;
-     size_t last_line;
-     size_t last_column;
+     int first_line;
+     int first_column;
+     int last_line;
+     int last_column;
 } aejson_loc_t;
 
 #define LOC_USER_ACTION \
